Creation of linked maps with outline
Spyder scripts executed by the jupyter notebook.
Diffusion maps and directed motion maps are created and linked. The entries of bins in the directed motion maps are deleted if at the same bin diffusion is present. Conversely, entries are kept in bins with diffusion showing preferential direction. These maps do include the outline of the structure which marks the region of special interest.