Generation outline marker for ROI
Spyder script. Removal of localizations belonging to the background is possible before the generation of the marker outline. Needs to be executed before the creation of the maps, but after the trajectory analysis. Final binning size, start and end of binning, as well as, padding parameter are needed to scale the outline onto the maps. Exemplary data in subfolder available.